requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
python-multipart==0.0.9
google-generativeai==0.5.0
elevenlabs==1.1.0
PyGithub==2.3.0
supabase==2.4.0
scikit-learn==1.4.0
numpy==1.26.4
Pillow==10.3.0
python-dotenv==1.0.1
requests==2.31.0